EightPuzzle hash method
Readded deleted hashCode method in EightPuzzle.java

So I wrote that long hash method before in the EightPuzzle class, as I
think in Java, elements of a Set need to override both equals() and
hashCode(). Since it didn't have a hashCode at first, the explored set
wasn't actually working, and this fixed it.