GridPuzzle is Operational!
So I've just made a GridPuzzle class that stores the grid. GridNode is
that state, and Direction (from Exercise 3) is the action. There is a
test maze that the algorithm can run through pretty easily.

There is also A GridPuzzleHeuristic which combines the cost and the
Manhattan distance between the current node and the goal (I haven't
tested it, but it should work).